refactor segment and add watershed
