+++ src/vistiq/segment/__init__.py
@@ -1,0 +1,92 @@
+"""Segmentation: thresholding, labeling, and region analysis."""
+
+from vistiq.segment._debug import debug_mask_labels
+from vistiq.segment.threshold import (
+    LocalThreshold,
+    LocalThresholdConfig,
+    NiblackThresholdConfig,
+    OtsuThreshold,
+    OtsuThresholdConfig,
+    RangeThreshold,
+    RangeThresholdConfig,
+    SauvolaThresholdConfig,
+    Thresholder,
+    ThresholderConfig,
+)
+from vistiq.segment.postprocess import (
+    BinaryProcessor,
+    BinaryProcessorConfig,
+    dilate_regions,
+    Watershed,
+    WatershedConfig,
+)
+from vistiq.segment.analysis import RegionAnalyzer, RegionAnalyzerConfig
+from vistiq.segment.select import (
+    RangeFilter,
+    RangeFilterConfig,
+    RegionFilter,
+    RegionFilterConfig,
+)
+from vistiq.segment.label import (
+    IterativeSegmenter,
+    IterativeSegmenterConfig,
+    LabelRemover,
+    LabelRemoverConfig,
+    Labeller,
+    LabellerConfig,
+    MicroSAMSegmenter,
+    MicroSAMSegmenterConfig,
+    Relabeler,
+    RelabelerConfig,
+    remap_dataframe_labels,
+    remap_labels,
+    remap_regionproperties,
+    remap_regions,
+    Segmenter,
+    SegmenterConfig,
+    SeriesSegmenter,
+    SeriesSegmenterConfig,
+)
+
+__all__ = [
+    "BinaryProcessor",
+    "BinaryProcessorConfig",
+    "Watershed",
+    "WatershedConfig",
+    "IterativeSegmenter",
+    "IterativeSegmenterConfig",
+    "LabelRemover",
+    "LabelRemoverConfig",
+    "Labeller",
+    "LabellerConfig",
+    "LocalThreshold",
+    "LocalThresholdConfig",
+    "MicroSAMSegmenter",
+    "MicroSAMSegmenterConfig",
+    "NiblackThresholdConfig",
+    "OtsuThreshold",
+    "OtsuThresholdConfig",
+    "RangeFilter",
+    "RangeFilterConfig",
+    "RangeThreshold",
+    "RangeThresholdConfig",
+    "RegionAnalyzer",
+    "RegionAnalyzerConfig",
+    "RegionFilter",
+    "RegionFilterConfig",
+    "Relabeler",
+    "RelabelerConfig",
+    "SauvolaThresholdConfig",
+    "Segmenter",
+    "SegmenterConfig",
+    "SeriesSegmenter",
+    "SeriesSegmenterConfig",
+    "Thresholder",
+    "ThresholderConfig",
+    "debug_mask_labels",
+    "dilate_regions",
+    "remap_dataframe_labels",
+    "remap_labels",
+    "remap_regionproperties",
+    "remap_regions",
+]
